+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased
 ----------
 - Added formatted message string to on_error parameters
 - Renamed project to 'py-buzz'
+- Added error sanitization for messages with embedded curly braces
 
 v0.1.7 - 2016-12-22
 -------------------